Fix ci pipeline and install process
