+++ .github/workflows/ci.yml
@@ -11,6 +11,11 @@ jobs:
       #----------------------------------------------
       - name: Check out repository
         uses: actions/checkout@v2
+      - name: Setup Julia
+        uses: julia-actions/setup-julia@v1
+        with:
+          version: 1
+          arch: x64
       - name: Set up python
         uses: actions/setup-python@v2
         with:
