+++ pyproject.toml
@@ -13,6 +13,10 @@ julia = "^0.5"
 pytest = "^6.2.5"
 pytest-cov = "^2.12.1"
 
+[tool.pytest.ini_options]
+addopts = "--cov=quickpomdps"
+
+
 [build-system]
 requires = ["poetry-core>=1.0.0"]
 build-backend = "poetry.core.masonry.api"
